analytics/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.37
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==42.0.8
